Extract cart token logic into trait, improve calculation precision, and streamline cookie management methods
